:recycle: Refactor legacy tx code
