+++ src/modules/Accounts/components/Item.js
@@ -4,16 +4,17 @@ import React, { useEffect, useRef } from 'react';
 import { View, TouchableOpacity } from 'react-native';
 import LottieView from 'lottie-react-native';
 import { translate } from 'react-i18next';
+import { useSelector } from 'react-redux';
+
+import { getTxConstant, isTransfer } from 'modules/Transactions/utils/helpers';
 import { fromRawLsk } from 'utilities/conversions';
 import { stringShortener } from 'utilities/helpers';
-import { getTxConstant, isTransfer } from 'modules/SendToken/constants';
 import loadingAnimation from 'assets/animations/loading-dots.json';
 import { Blur } from 'components/shared/blur';
 import FormattedNumber from 'components/shared/formattedNumber';
 import FormattedDate from 'components/shared/formattedDate';
 import { B, Small } from 'components/shared/toolBox/typography';
 import withTheme from 'components/shared/withTheme';
-import { useSelector } from 'react-redux';
 import Symbol from './Symbol';
 import getStyles from './styles';
