+++ src/modules/SendToken/__fixtures__/index.js
@@ -1,3 +1,1 @@
-export * from './mockTokensMeta';
-export * from './mockFeesByPriorities';
 export * from './mockSendTokenResponse';
