--- src/modules/SendToken/constants/index.js
@@ -1,2 +1,0 @@
-export * from './transactions';
-export * from './constants';
