+++ src/modules/Transactions/utils/helpers.js
@@ -1,44 +1,11 @@
 import * as Lisk from '@liskhq/lisk-client';
+import i18next from 'i18next';
 
-// TODO: Use from service endpoint/elements
-export const baseTransactionSchema = {
-  $id: '/lisk/baseTransaction',
-  type: 'object',
-  required: ['module', 'command', 'nonce', 'fee', 'senderPublicKey', 'params'],
-  properties: {
-    module: {
-      dataType: 'string',
-      fieldNumber: 1,
-    },
-    command: {
-      dataType: 'string',
-      fieldNumber: 2,
-    },
-    nonce: {
-      dataType: 'uint64',
-      fieldNumber: 3,
-    },
-    fee: {
-      dataType: 'uint64',
-      fieldNumber: 4,
-    },
-    senderPublicKey: {
-      dataType: 'bytes',
-      fieldNumber: 5,
-    },
-    params: {
-      dataType: 'bytes',
-      fieldNumber: 6,
-    },
-    signatures: {
-      type: 'array',
-      items: {
-        dataType: 'bytes',
-      },
-      fieldNumber: 7,
-    },
-  },
-};
+import { themes } from 'constants/styleGuide';
+import txUnknownLight from 'assets/images/txDetail/tx-unknown-light.png';
+import txUnknownDark from 'assets/images/txDetail/tx-unknown-dark.png';
+
+import { BASE_TRANSACTION_SCHEMA, MODULE_COMMAND_MAP, TRANSACTIONS } from './constants';
 
 export const getCommandParamsSchema = (module, command, schema) => {
   const moduleCommand = module.concat(':', command);
@@ -53,7 +20,7 @@ export const getCommandParamsSchema = (module, command, schema) => {
 };
 
 export const decodeBaseTransaction = (encodedTransaction) =>
-  Lisk.codec.codec.decode(baseTransactionSchema, encodedTransaction);
+  Lisk.codec.codec.decode(BASE_TRANSACTION_SCHEMA, encodedTransaction);
 
 export const decodeTransaction = (encodedTransaction, paramsSchema) => {
   const transaction = decodeBaseTransaction(encodedTransaction);
@@ -77,7 +44,7 @@ export const encodeTransaction = (transaction, paramsSchema) => {
     encodedParams = transaction.params;
   }
 
-  const decodedTransaction = Lisk.codec.codec.encode(baseTransactionSchema, {
+  const decodedTransaction = Lisk.codec.codec.encode(BASE_TRANSACTION_SCHEMA, {
     ...transaction,
     params: encodedParams,
   });
@@ -86,7 +53,7 @@ export const encodeTransaction = (transaction, paramsSchema) => {
 };
 
 export const fromTransactionJSON = (transaction, paramsSchema) => {
-  const tx = Lisk.codec.codec.fromJSON(baseTransactionSchema, {
+  const tx = Lisk.codec.codec.fromJSON(BASE_TRANSACTION_SCHEMA, {
     ...transaction,
     params: '',
   });
@@ -111,13 +78,13 @@ export const fromTransactionJSON = (transaction, paramsSchema) => {
 export const toTransactionJSON = (transaction, paramsSchema) => {
   if (Buffer.isBuffer(transaction.params)) {
     return {
-      ...Lisk.codec.codec.toJSON(baseTransactionSchema, transaction),
+      ...Lisk.codec.codec.toJSON(BASE_TRANSACTION_SCHEMA, transaction),
       params: paramsSchema ? Lisk.codec.codec.decodeJSON(paramsSchema, transaction.params) : {},
       id: transaction.id.toString('hex'),
     };
   }
   return {
-    ...Lisk.codec.codec.toJSON(baseTransactionSchema, {
+    ...Lisk.codec.codec.toJSON(BASE_TRANSACTION_SCHEMA, {
       ...transaction,
       params: Buffer.alloc(0),
     }),
@@ -125,3 +92,16 @@ export const toTransactionJSON = (transaction, paramsSchema) => {
     id: transaction.id && transaction.id.toString('hex'),
   };
 };
+
+export const getTxConstant = ({ moduleAssetId }) => {
+  const result = TRANSACTIONS[moduleAssetId] ?? {};
+  return {
+    ...result,
+    title: result?.title ?? i18next.t('Transaction'),
+    image: result?.image
+      ? result?.image
+      : (theme) => (theme === themes.light ? txUnknownLight : txUnknownDark),
+  };
+};
+
+export const isTransfer = ({ moduleAssetId }) => moduleAssetId === MODULE_COMMAND_MAP.transfer;
